add config file v1
