+++ src/main/java/AppByDataBaseStorage.java
@@ -1,0 +1,4 @@
+public class AppByDataBaseStorage {
+    static void start() {
+    };
+}
