+++ src/main/java/ConfigReader.java
@@ -1,0 +1,15 @@
+import java.nio.file.Path;
+import java.util.Properties;
+
+public class ConfigReader {
+    Path path;
+
+    public ConfigReader(Path path) {
+        this.path = path;
+    }
+
+    Properties properties = new Properties();
+    public DbType parse() {
+        return DbType.STORAGE_JSON;
+    }
+}
