+++ src/main/java/DataSourceManager.java
@@ -1,0 +1,17 @@
+public class DataSourceManager {
+    ConfigReader configReader;
+
+    public DataSourceManager(ConfigReader configReader) {
+        this.configReader = configReader;
+    }
+
+    void start (){
+    /*    DbType dbType = configReader.parse();
+        if (dbType.name().equals("STORAGE_JSON")){*/
+            AppByJsonStorage.start();
+       /* }
+        if (dbType.name().equals("STORAGE_DATABASE")){
+            AppByDataBaseStorage.start();
+        }*/
+    };
+}
