+++ src/main/java/DbType.java
@@ -1,0 +1,3 @@
+public enum DbType {
+    STORAGE_JSON, STORAGE_DATABASE
+}
