+++ src/main/java/runner/DataBaseModeRunner.java
@@ -1,0 +1,8 @@
+package runner;
+
+public class DataBaseModeRunner implements ModeRunner {
+    @Override
+    public void start() {
+
+    }
+}
