+++ src/main/resources/app.properties
@@ -1,0 +1,1 @@
+DB_TYPE=Storage_json
