--- src/main/resources/test.json
@@ -1,13 +1,0 @@
-[ {
-  "name" : "beer",
-  "price" : 50.0,
-  "quantity" : 30
-}, {
-  "name" : "cola",
-  "price" : 20.0,
-  "quantity" : 20
-}, {
-  "name" : "soap",
-  "price" : 30.0,
-  "quantity" : 10
-} ]
